refactor(execution): 重构执行器中的比较和类型转换逻辑
- 将比较和类型转换相关的函数移至 Value 类中
- 优化 GroupExecutor 中的分组逻辑
- 调整 IndexScanExecutor 中的比较操作
- 移除不必要的调试输出
